Fix some tests. Refactor.
