+++ tracer/test/test-applications/integrations/Samples.InstrumentedTests/Vulnerabilities/SqlInjection/DabaseHelpers/ApplicationDbContextCore.cs
@@ -1,7 +1,6 @@
 #if NETCOREAPP3_0_OR_GREATER
 
 using Microsoft.EntityFrameworkCore;
-using System;
 using System.Data.Common;
 
 namespace Samples.InstrumentedTests.Iast.Vulnerabilities.SqlInjection;
